backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# AI & LLM
langgraph==0.3.34
langchain==0.3.27
langchain-anthropic==0.3.35
anthropic>=0.40.0

# Okta Authentication
okta-jwt-verifier>=0.2.3
python-jose[cryptography]==3.3.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Auth0 / FGA
openfga-sdk==0.9.7

# HTTP & Networking
httpx==0.27.0
websockets==12.0

# Utilities
pydantic>=2.5.0
python-dotenv==1.0.0

# Token encoding (for context window management)
tiktoken>=0.5.0
